// src/types/application.type.ts
export interface IApplicationPayload {
  name: string
  email: string
  phone: string
  company_name: string
  message: string
}

export interface IApplicationForm {
  name: string
  email: string
  phone: string
  company_name: string
  website: string
  products: string
  labTesting: string
  message: string
}